Move `SeedIsGarden`
This move is necessary to please the import restrictions of`pkg/extensions`.
