+++ skaffold.yaml
@@ -962,6 +962,7 @@ build:
         - pkg/utils/errors
         - pkg/utils/flow
         - pkg/utils/gardener
+        - pkg/utils/gardener/operator
         - pkg/utils/gardener/secretsrotation
         - pkg/utils/gardener/shootstate
         - pkg/utils/gardener/tokenrequest
